pyproject.toml
[tool.poetry]
name = "store-tracker"
version = "0.1.0"
description = "Django-based app for comparing prices between different stores."
authors = ["Nikita Solouyou <solovev.nikita.05@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
djoser = "2.2.0"
django-environ = "0.10.0"
django-filter = "23.2"
django-cleanup = "7.0.0"
django-redis = "5.2.0"
django-cors-headers = "4.2.0"
celery = "5.2.7"
pillow = "9.5.0"
eventlet = "0.33.3"
gunicorn = "20.1.0"
drf-spectacular = "^0.26.4"
psycopg2-binary = "^2.9.7"
django-cacheops = "^7.0.1"
django-silk = "^5.0.3"
pydantic = "^2.4.2"

[tool.poetry.group.dev.dependencies]
mypy = "1.4.1"
django-stubs = {version = "4.2.3", extras = ["compatible-mypy"]}
djangorestframework-stubs = {version = "3.14.2", extras = ["compatible-mypy"]}
mixer = "7.2.2"
faker = "12.0.1"
pytest-django = "4.5.2"
pre-commit = "^3.3.3"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.287"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
filterwarnings = [
    "ignore::DeprecationWarning:kombu.*:0",
    "ignore::DeprecationWarning:celery.*:0",
    "ignore::DeprecationWarning:rest_framework_simplejwt.*:0",
]

[tool.ruff]
line-length = 119
exclude = [
    "settings",
    "migrations",
]

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
    "pydantic.mypy",
]

ignore_missing_imports = "true"

[tool.django-stubs]
django_settings_module = "core.settings.local"
